feat: optimize file structure for parse map and do some norm checks
