+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: jakoh <jakoh@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/09 14:42:20 by jakoh             #+#    #+#             */
-/*   Updated: 2023/09/11 16:46:38 by jakoh            ###   ########.fr       */
+/*   Updated: 2023/09/11 17:04:25 by jakoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,6 @@ int	main(int ac, char **av)
 		printf("%s\n", variables.map.map[i]);
 	}
 	free_map(variables.map.map);
-	// system("leaks cub3d");
+	system("leaks cub3d");
 	return (2);
 }
-
